error fixes
