+++ fly.toml
@@ -1,0 +1,46 @@
+# fly.toml file generated for guess-the-spanish-word on 2022-09-15T01:02:48-04:00
+
+app = "guess-the-spanish-word"
+kill_signal = "SIGINT"
+kill_timeout = 5
+processes = []
+
+[build]
+  builder = "heroku/buildpacks:20"
+
+[env]
+  PORT = "8080"
+  REACT_APP_PORT="8080"
+  ATLAS_URI = "mongodb+srv://User1:User1@cluster0.unlledw.mongodb.net/?retryWrites=true&w=majority"
+  REACT_APP_SERVER_URL = "https://guess-the-spanish-word.fly.dev"
+
+
+[experimental]
+  allowed_public_ports = []
+  auto_rollback = true
+
+[[services]]
+  http_checks = []
+  internal_port = 8080
+  processes = ["app"]
+  protocol = "tcp"
+  script_checks = []
+  [services.concurrency]
+    hard_limit = 25
+    soft_limit = 20
+    type = "connections"
+
+  [[services.ports]]
+    force_https = true
+    handlers = ["http"]
+    port = 80
+
+  [[services.ports]]
+    handlers = ["tls", "http"]
+    port = 443
+
+  [[services.tcp_checks]]
+    grace_period = "1s"
+    interval = "15s"
+    restart_limit = 0
+    timeout = "2s"
