+++ package.json
@@ -1,4 +1,5 @@
 {
+  "homepage": ".",
   "name": "guess-the-word",
   "version": "0.1.0",
   "private": true,
